requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
decord==0.6.0
et-xmlfile==2.0.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
hf-xet==1.1.9
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
jinja2==3.1.6
kiwisolver==1.4.9
markupsafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
portalocker==3.2.0
pycocotools==2.0.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
safetensors==0.6.2
schedulefree==1.4.1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
timm==1.0.19
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
